requirements.txt
PyMuPDF
langchain-community
langchain-huggingface
python-dotenv
google-genai

langchain 
langchain-core
langchain-google-genai
google-generativeai
chromadb